--- a/benchmark/clickbench/q29.benchmark
+++ b/benchmark/clickbench/q29.benchmark
@@ -5,3 +5,4 @@
 template benchmark/clickbench/clickbench-noresult.benchmark.in
 QUERY_NUMBER=29
 QUERY_NUMBER_PADDED=29
+RESULT_QUERY=SELECT k FROM __answer