--- a/src/common/assert.cpp
+++ b/src/common/assert.cpp
@@ -7,7 +7,8 @@ void DuckDBAssertInternal(bool condition, const char *condition_name, const char
 	if (condition) {
 		return;
 	}
-	throw InternalException("Assertion triggered in file \"%s\" on line %d: %s", file, linenr, condition_name);
+	throw InternalException("Assertion triggered in file \"%s\" on line %d: %s%s", file, linenr, condition_name,
+	                        Exception::GetStackTrace());
 }
 
 } // namespace duckdb