--- a/src/common/exception.cpp
+++ b/src/common/exception.cpp
@@ -9,6 +9,9 @@
 #include <stdio.h>
 #include <stdlib.h>
 #endif
+#ifdef DUCKDB_DEBUG_STACKTRACE
+#include <execinfo.h>
+#endif
 
 namespace duckdb {
 
@@ -34,6 +37,24 @@ bool Exception::UncaughtException() {
 	return std::uncaught_exceptions() > 0;
 #else
 	return std::uncaught_exception();
+#endif
+}
+
+string Exception::GetStackTrace(int max_depth) {
+#ifdef DUCKDB_DEBUG_STACKTRACE
+	string result;
+	auto callstack = unique_ptr<void *[]>(new void *[max_depth]);
+	int frames = backtrace(callstack.get(), max_depth);
+	char **strs = backtrace_symbols(callstack.get(), frames);
+	for (int i = 0; i < frames; i++) {
+		result += strs[i];
+		result += "\n";
+	}
+	free(strs);
+	return "\n" + result;
+#else
+	// Stack trace not available. Toggle DUCKDB_DEBUG_STACKTRACE in exception.cpp to enable stack traces.
+	return "";
 #endif
 }