--- a/src/execution/physical_plan/plan_create_table.cpp
+++ b/src/execution/physical_plan/plan_create_table.cpp
@@ -1,8 +1,8 @@
 #include "duckdb/catalog/catalog_entry/scalar_function_catalog_entry.hpp"
 #include "duckdb/execution/operator/schema/physical_create_table.hpp"
-#include "duckdb/execution/operator/schema/physical_create_table_as.hpp"
 #include "duckdb/execution/physical_plan_generator.hpp"
 #include "duckdb/parser/parsed_data/create_table_info.hpp"
+#include "duckdb/execution/operator/persistent/physical_insert.hpp"
 #include "duckdb/planner/expression/bound_function_expression.hpp"
 #include "duckdb/planner/expression_iterator.hpp"
 #include "duckdb/planner/operator/logical_create_table.hpp"
@@ -33,7 +33,7 @@ unique_ptr<PhysicalOperator> PhysicalPlanGenerator::CreatePlan(LogicalCreateTabl
 	bool replace = op.info->Base().on_conflict == OnCreateConflict::REPLACE_ON_CONFLICT;
 	if ((!existing_entry || replace) && !op.children.empty()) {
 		D_ASSERT(op.children.size() == 1);
-		auto create = make_unique<PhysicalCreateTableAs>(op, op.schema, move(op.info), op.estimated_cardinality);
+		auto create = make_unique<PhysicalInsert>(op, op.schema, move(op.info), op.estimated_cardinality);
 		auto plan = CreatePlan(*op.children[0]);
 		create->children.push_back(move(plan));
 		return move(create);